* Divided the unit tests and integration tests into two separate source files, 'test_integ_Packager.py' and 'test_unit_Packager.py', respectively.
* Modified 'tools/unit_tests_and_wheel_creation.sh' to run pytest.

* Renamed dir 'src/son/package/tests/.son-schema' to 'src/son/package/tests/son-schema'.
