+++ src/son/package/tests/son-schema/vnfd-schema.yml
